+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/statement/ddl/DDLStatementAssert.java
@@ -20,6 +20,10 @@ package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.SQLCaseAssertContext;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.impl.AlterTableStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.impl.CreateIndexStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.impl.CreateTableStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.impl.DropIndexStatementAssert;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.SQLParserTestCase;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.ddl.AlterTableStatementTestCase;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.ddl.CreateIndexStatementTestCase;
