+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/statement/ddl/impl/CreateTableStatementAssert.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.impl;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
