+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/statement/ddl/impl/DropIndexStatementAssert.java
@@ -15,44 +15,47 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.impl;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.SQLCaseAssertContext;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.table.TableAssert;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.ddl.CreateIndexStatementTestCase;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.ddl.DropIndexStatementTestCase;
 import org.apache.shardingsphere.sql.parser.sql.segment.generic.TableSegment;
-import org.apache.shardingsphere.sql.parser.sql.statement.ddl.CreateIndexStatement;
-
-import java.util.Collections;
+import org.apache.shardingsphere.sql.parser.sql.statement.ddl.DropIndexStatement;
 
 /**
- * Create index statement assert.
+ * Drop index statement assert.
  *
  * @author zhangliang
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class CreateIndexStatementAssert {
+public final class DropIndexStatementAssert {
     
     /**
-     * Assert create index statement is correct with expected parser result.
+     * Assert drop index statement is correct with expected parser result.
      * 
      * @param assertContext Assert context
-     * @param actual actual create index statement
-     * @param expected expected create index statement test case
+     * @param actual actual drop index statement
+     * @param expected expected drop index statement test case
      */
-    public static void assertIs(final SQLCaseAssertContext assertContext, final CreateIndexStatement actual, final CreateIndexStatementTestCase expected) {
+    public static void assertIs(final SQLCaseAssertContext assertContext, final DropIndexStatement actual, final DropIndexStatementTestCase expected) {
         assertTable(assertContext, actual, expected);
         assertIndex(assertContext, actual, expected);
     }
     
-    private static void assertTable(final SQLCaseAssertContext assertContext, final CreateIndexStatement actual, final CreateIndexStatementTestCase expected) {
-        TableAssert.assertIs(assertContext, actual.findSQLSegments(TableSegment.class), Collections.singletonList(expected.getTable()));
+    private static void assertTable(final SQLCaseAssertContext assertContext, final DropIndexStatement actual, final DropIndexStatementTestCase expected) {
+        TableAssert.assertIs(assertContext, actual.findSQLSegments(TableSegment.class), expected.getTables());
     }
     
-    private static void assertIndex(final SQLCaseAssertContext assertContext, final CreateIndexStatement actual, final CreateIndexStatementTestCase expected) {
+    private static void assertIndex(final SQLCaseAssertContext assertContext, final DropIndexStatement actual, final DropIndexStatementTestCase expected) {
         // TODO should assert index for all databases(mysql and sqlserver do not parse index right now)
-//        IndexAssert.assertIs(assertContext, actual.getIndex(), expected.getIndex());
+//        assertThat(assertContext.getText("Index size assertion error: "), actual.getIndexes().size(), is(expected.getIndexes().size()));
+//        int count = 0;
+//        for (IndexSegment each : actual.getIndexes()) {
+//            IndexAssert.assertIs(assertContext, each, expected.getIndexes().get(count));
+//            count++;
+//        }
     }
 }
