+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/statement/dml/impl/InsertStatementAssert.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.dml;
+package org.apache.shardingsphere.sql.parser.integrate.asserts.statement.dml.impl;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
